refactor: CJB-17 컴포넌트 네이밍 PascalCase로 개선
