+++ app/festival/[id]/_modules/MainImage.tsx
